Update key_creation
Finalised the key_creation script. The main aspect of the script is in the if-else section. AS of now, the script will output a 16x11 matrix which contains all-round keys (10) and the original key (1). A future job should include creating another function to reduce the 50 line code to a readable format (≈ 20).