Updated user_input_decrypt
Changed the script name from decrypt_user_input to user_input_decrypt as previously based on the name alone one would assume that script did a lot more than it actually did.

I've also added a function which takes the user input (from a dialog box) and format that to a 16 block matrix.