Updated app interface
Passed down the information the user enters in the GUI text box to main_encrypt. This meant main_encrypt had to change to a function  with an intake argument "userData" which is passed down to user_input which formats this data. The output of main_encrypt is then passed back to the GUI and displayed in the text box. The same was done for main_decrypt and user_input_decrypt.