Changed two script names
I've changed the names of user_input_encrypt to format_encrypt_in and user_input_decrypt to format_decrypt_in, this change has been updated to the function calls in main_encrypt and main_decrypt. The reason for the change in names was due to comfusion arising to the functinality of the script, as they were now formating input instead of taking users input.