Created key check
I've created a new script called key_check which checks the entered key. If the entered key doesn't pass, an error dialog box will be displayed informing the user to check the key. I've also added a hover-over text on 'Key' which shows the user examples on the expect style of keys.