Implimented 192-bit key creation
I've added/change certain factors of the key_creation script. Now it can correctly create both 128-bit keys and 192-bit keys. To achieve this an extra argument was required to be passed through to the function called "AESMode". This can either have a value "128-bit" or "192-bit". Once determined the script will create the required key.