CBC on main_encrypt
Added an extra argument to main_encryp, "AESType" which can either be "ECB" or "CBC". If CBC then the added if statement will be executed. This if statement XORs the encrypted cipher block with the future (next block) plaintext block, hence making all the blocks dependent on the previous block (except block 1, which needs IV to be implemented).