IV in encrypt/decrypt scripts
I've added the implications of IV value to CBC mode on main_encrypt and main_decrypt. This was relatively straight forward as all it required in main_encrypt is the XOR of IV value and first plaintext block. For main_decrypt the XOR between first decrypted cipher block and IV value.