New script gf_conv.m
Created a new script called gf_conv. This function does exactly the same function as gfconv(value1, value2, 2). It is the multiplication of two 8 bit values in GF(2). The reason for this script is to make avoid the use of high-level inbuilt MATLAB functions, as transferring, this information to Verilog would make it more difficult. I've also added a help icon on the GUI which contains information on different sections of fields/text boxes.