Implimented data format
I've implemented a new feature that allows the user to encrypt hex data (instead of only Base64). This feature also allows the decryption of the data to stay in hex (not converted to char).

This required a new select box in the GUI (Data format input/output), which had to be passed on main_encrypt and main_decrypt to perform the wanted action